+++ src/core/network/network_builder.hpp
@@ -6,6 +6,7 @@
 #include "core/encoding/rate_encoder.hpp"
 #include "core/neural/neuron.hpp"
 #include "core/neural/neuron_pool.hpp"
+#include "core/plasticity/homeostasis.hpp"
 #include "core/plasticity/stdp.hpp"
 #include "core/spatial/lattice.hpp"
 #include "core/spatial/neighbor_index.hpp"
@@ -29,7 +30,7 @@ struct NetworkConfig {
   uint64_t seed = 42;
   neural::LIFParams lif_params{};
   synaptic::SynapseParams synapse_params{};
-  temporal::HomeostasisConfig homeostasis{};
+  plasticity::HomeostasisConfig homeostasis{};
   encoding::RateEncoderParams encoder_params{};
   float decoder_window_ms = 50.0f;
   plasticity::STDPParams stdp_params{};